Reduce memory usage for the pod logs endpoint (#5178)
* Stream pod logs while processing

The pod logs endpoint was loading the pod logs on memory, then parsing/process them (also on memory) and, finally sending the results to the browser/client. This is leading to high memory usage for large logs.

This is changing the approach to parse/process logs and send results to the client/browser while also reading, reducing memory usage to a minimum.

Additionally, a bug was solved when SinceTime and Duration params are both set: depending on available logs, it was possible that the returned logs didn't belong to the requeted time window.

* Replace tailLines with maxLines

The <tailLines dropdown plus endTime> and the streaming process are not "very good friends". If the user chooses a "Custom" time range and chooses to specify an "end time" and ALSO specifies to tail some lines rather than "all lines", then "streaming" cannot be done in the back-end and two different approaches were needed to feed logs to the front-end (one where streaming was possible, and another one without streaming). To avoid having two different approaches to collect logs, the tailLines option was removed. Moreover, the tailLines is partially redundant with the duration and/or startTime selectors because there is a point where it does not make sense to increase the duration if tailLines is also not increased.

The tailLines was replaced with a similar "maxLines" option. The difference is that rather than showing the "last N lines" for the chosen time range, it will show the "first N lines" for the selected time range. A warning is shown to the user if the chosen maximum number of lines is reached to tell that logs are truncated.

Co-authored-by: Jay Shaughnessy <jshaughn@users.noreply.github.com>