Metrics tab: Ability to show both reporters (#5284)
* Back-end: Allow returning metrics for both reporters

This removes the constraint that a reporter is required in the back-end for querying metrics. To return metrics for both reporters, the URL argument should be reporter=both which will generate a prometheus query without the reporter label. This would, of course, aggregate values for both reporters in a single timeserie. In order to have them properly separated, the client should include a byLbl[]=reporter parameter in the URL, which is already an existent/supported parameter.

* Front-end: Add support for showing both reporters in metrics tabs

* Add a "both" option in the "Reporter" drop-down
* Correctly pass the needed parameters to the back-end API when both reporters are requested.
* Some initial series pairing when both reporters are requested. This is in preparation for showing an area chart rather than a line chart. Hopefully, this would allow the user to better correlate differences between "same" timeseries where only the reporter changes.

* Fix error when there is no data on one of the reporters

* Show source and destination values in tooltip

... when "both" reporters are enabled.

* Use average values for calculating trend-lines for "both" reporters

* Consider for unsynced series