Launch wizards from graph's side panel (#5323)
* Launch wizards from graph's side panel and context menu

Some chunks of code were moved to new components to reduce code duplications as much as possible. A custom react hook was needed.

At first, GraphPage.tsx was planned to store the service details on its state to share it between the side panel and the graph context menu. However, the context menu can be open for a different node than the one for the side panel which means that sharing data does not make sense. The GraphPage.tsx is limited to only showing the wizard or showing the delete confirm dialog.