Fix refresh interval in Kiosk mode (#5433)
* Fix refresh interval in Kiosk mode

The Refresh.tsx component (and redux) was being used for refresh logic. Since this is the component that is added in pages that need refresh data periodically (and thus, need the duration/time-interval and refresh interval controls), it made sense that this component also fired the refreshes. But in Kiosk mode this component is not rendered, thus refreshes are not being fired.

To fix the refresh for Kiosk mode, the timer has to be removed from the Refresh.tsx component, and also Redux is no longer involved for triggering refreshes (although Redux is still involved for storing/updating the desired interval). A useRefreshInterval custom react-hook is created to host the timer and this is the hook that function components should use to watch/subscribe for refreshes. For class components, the new RefreshNotifier.tsx component exposes similar functionality.

The refresh timer will be active while there are active "subscriptions" and the auto-refresh is not "paused". Since Redux is no longer involved in refreshes, a triggerRefresh function is also created to programmatically trigger a refresh (i.e. for refresh buttons).

Since several components were relying on Redux to watch/listen when a refresh is triggered, a connectRefresh.tsx HOC is also created to provide a similar interface, and mainly to reduce the churn of the changes required for the fix. However, it is probably better to use the useRefreshInterval hook or the RefreshNotifier.tsx component because they make explicit to "clients" when a refresh is happening, rather than trying to determine the refresh event in the componentDidUpdate method.

Related kiali/openshift-servicemesh-plugin#42

* Fix build error, because of rebasing

* Fix tests

* Fix invalid ref

* Fix rebase build issue

* Add comments to the RefreshNotifier.tsx component