Initial PFT Graph Impl (#6093)
The PFT graph is hidden behind an undocumented feature flag, spec.kiali_feature_flags.ui_defaults.graph.impl. If if set to 'pf' it will be the graph. If set to 'both' then both graph impls will be available (useful for comparison when developing).

It is important that future graph work consider both the Cytoscape and Patternfly impls and that code be shared if possible.

A Few Notes:
* edge.tsx, node.tsx, group.tsx, groupCollapsed.tsx, groupExpanded.tsx are copies of PFT components that are then customized for things we need in Kiali.  We should work to move those customizations into PFT upstream, or find alternatives that allow us to get away from this approach, as the copied components are "frozen" to the PFT version we copied from, and will not pick up upstream changes.
* note that to get edges/nodes to re-render you need to make a formal change (i.e. call setData() as opposed to just updating the data manually.  PFT renders only what it needs, not everything top-down (meaning the TopologyView may render itself, but not re-render each node or edge it contains).
* Store graphData under controller.getGraph.getData(), this is akin to our globalScratchData for Cy.
* PFT has a much simpler animation with just 5 different speeds.  We adapt our more complex mechanism into
the PFT model by breaking our "delay" times into 5 buckets.